+++ fixtures/ext-types/uniffi-one/src/lib.rs
@@ -1,0 +1,5 @@
+pub struct UniffiOneType {
+    pub sval: String,
+}
+
+include!(concat!(env!("OUT_DIR"), "/uniffi-one.uniffi.rs"));