+++ uniffi_bindgen/src/bindings/ruby/gen_ruby.rs
@@ -190,6 +190,8 @@ mod filters {
                     )
                 }
             }
+            Type::External { .. } => panic!("No support for external types, yet"),
+            Type::Wrapped { .. } => panic!("No support for wrapped types, yet"),
         })
     }
 
@@ -221,6 +223,8 @@ mod filters {
                 class_name_rb(&type_.canonical_name())?,
                 nm
             ),
+            Type::External { .. } => panic!("No support for lowering external types, yet"),
+            Type::Wrapped { .. } => panic!("No support for lowering wrapped types, yet"),
         })
     }
 
@@ -251,6 +255,8 @@ mod filters {
                 nm,
                 class_name_rb(&type_.canonical_name())?
             ),
+            Type::External { .. } => panic!("No support for lifting external types, yet"),
+            Type::Wrapped { .. } => panic!("No support for lifting wrapped types, yet"),
         })
     }
 }