+++ uniffi_bindgen/src/bindings/swift/gen_swift.rs
@@ -190,6 +190,8 @@ mod filters {
             Type::Optional(type_) => format!("{}?", type_swift(type_)?),
             Type::Sequence(type_) => format!("[{}]", type_swift(type_)?),
             Type::Map(type_) => format!("[String:{}]", type_swift(type_)?),
+            Type::External { .. } => panic!("No support for lifting types, yet"),
+            Type::Wrapped { .. } => panic!("No support for lifting types, yet"),
         })
     }